Refactor: ProfileSection Hook으로 분리
